#6 map_fragment/지역,지하철역,건물 검색 화면 : domain 레이어 오타로 인한 리팩토링 🔨  - entitiy > entity 로 이동  - OnboardingResultViewModel > SearchCompanyResultViewModel로 rename  - SearchResult > SearchCompanyResult로 rename
