+++ app/src/main/java/com/ftw/hometerview/ui/searchcompanyresult/SearchCompanyResultViewModel.kt
@@ -3,8 +3,8 @@ package com.ftw.hometerview.ui.searchcompanyresult
 import android.util.Log
 import com.ftw.hometerview.BR
 import androidx.lifecycle.ViewModel
-import com.ftw.domain.entitiy.DEMO_RESULTS
-import com.ftw.domain.entitiy.SearchResult
+import com.ftw.domain.entity.DEMO_COMPANY_RESULTS
+import com.ftw.domain.entity.SearchCompanyResult
 import com.ftw.hometerview.R
 import com.ftw.hometerview.adapter.RecyclerItem
 import com.ftw.hometerview.dispatcher.Dispatcher
@@ -16,7 +16,7 @@ import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
 
-class OnboardingResultViewModel(
+class SearchCompanyResultViewModel(
     private val dispatcher: Dispatcher
 ) : ViewModel() {
 
@@ -41,13 +41,13 @@ class OnboardingResultViewModel(
 
     }
 
-    private fun createSearchResultRecycleModel(searchResult: SearchResult): SearchResultRecycleModel {
+    private fun createSearchResultRecycleModel(searchResult: SearchCompanyResult): SearchResultRecycleModel {
         return SearchResultRecycleModel(searchResult).apply {
             itemClickListener = { searchResult -> showClickMessage(searchResult) }
         }
     }
 
-    private fun showClickMessage(searchResult: SearchResult) {
+    private fun showClickMessage(searchResult: SearchCompanyResult) {
         //통신으로 받아온 정보를 State와 조합해서 사용하면 될 듯 함
 
     }
@@ -70,7 +70,7 @@ class OnboardingResultViewModel(
         }
     }
 
-    private fun loadDemoDatas(): List<SearchResult> = DEMO_RESULTS
+    private fun loadDemoDatas(): List<SearchCompanyResult> = DEMO_COMPANY_RESULTS
 
 }
