+++ app/src/main/java/com/ftw/hometerview/ui/searchcompanyresult/recyclemodel/SearchResultRecycleModel.kt
@@ -1,10 +1,10 @@
 package com.ftw.hometerview.ui.searchcompanyresult.recyclemodel
 
-import com.ftw.domain.entitiy.SearchResult
+import com.ftw.domain.entity.SearchCompanyResult
 
-class SearchResultRecycleModel(val result: SearchResult) {
+class SearchResultRecycleModel(val result: SearchCompanyResult) {
 
-    lateinit var itemClickListener: (result: SearchResult) -> Unit?
+    lateinit var itemClickListener: (result: SearchCompanyResult) -> Unit?
 
     fun onItemClick() {
         itemClickListener(result)
