+++ domain/src/main/java/com/ftw/domain/entity/BuildingMarker.kt
@@ -1,4 +1,4 @@
-package com.ftw.domain.entitiy
+package com.ftw.domain.entity
 
 data class BuildingMarker(
     val buildingName: String,
